+++ srcs/objects/cylinder.c
@@ -1,6 +1,6 @@
 #include "miniRT.h"
 
-float	cylinder_intersection(t_minirt *minirt, t_ray ray, t_object cylinder)
+double	cylinder_intersection(t_minirt *minirt, t_ray ray, t_object cylinder)
 {
 		(void)minirt;	
 	(void)ray;
